[WIP] Update FFT base types

+++ waveforms.cpp
@@ -8,7 +8,7 @@ void Waveform::ramp(uint16_t *sink,  unsigned int points, double min, double max
   float last_val = 0.0;
   float step_val = ((float)max / (float)((float)points / 2.0));
 
-  for(unsigned short int i = min; i < (points + min); i++) {
+  for(unsigned int i = min; i < (points + min); i++) {
     if((unsigned int)(last_val + step_val) > max || (unsigned int)(last_val + step_val) < min) {
       step_val *= -1.0F;
     }
